+++ fltk-derive/src/group.rs
@@ -70,17 +70,16 @@ pub fn impl_group_trait(ast: &DeriveInput) -> TokenStream {
                 unsafe {
                     assert!(!self.was_deleted());
                     let sz = self.children();
-                    let mut v: Vec<Widget> = vec![];
+                    let mut v: Vec<Option<Box<dyn FnMut()>>> = vec![];
                     if sz > 0 {
                         for i in 0..sz {
-                            v.push(self.child(i).unwrap());
+                            v.push(self.child(i).unwrap().user_data());
+                            v.push(self.child(i).unwrap().draw_data());
                         }
                     }
                     #clear(self._inner);
-                    if sz > 0 {
-                        for mut child in v {
-                            child.cleanup();
-                        }
+                    if let Some(mut p) = self.parent() {
+                        p.redraw();
                     }
                 }
             }